+++ CHANGELOG.md
@@ -13,3 +13,4 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Chemical support in the Grid.
 - Chemical support in the Storage Monitor.
 - Chemical support in the Importer.
+- Chemical support in the Exporter.